+++ packages/statemanager/test/stateManager.storage.spec.ts
@@ -14,11 +14,10 @@ import { Caches, MerkleStateManager } from '../src/index.ts'
 
 import { createAccountWithDefaults } from './util.ts'
 
-const isBrowser = new Function('try {return this===window;}catch(e){ return false;}')
 describe('StateManager -> Storage', () => {
   for (const storageCacheOpts of [{ size: 1000 }, { size: 0 }]) {
     for (const prefixStorageTrieKeys of [false, true]) {
-      it.skipIf(isBrowser() === true)(`should dump storage`, async () => {
+      it(`should dump storage`, async () => {
         const stateManager = new MerkleStateManager({
           prefixStorageTrieKeys,
           caches: new Caches({ storage: storageCacheOpts }),